requirements.txt
discord.py==2.3.2
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==21.2.0 